chore(rest): move to root of lib
